Move the basic_common_reference specializations into List.h, add a real guard
List_inl.h is included from List.h's own last line and itself includes
ivalue.h, which cycles back into List.h as a #pragma once no-op. Under
that ordering, anything ivalue.h/ivalue_inl.h/Dict.h/jit_type.h pulls
in got preprocessed while ListElementReference was complete but the
specialization not yet declared -- IFNDR, not a live bug today, but a
silent trap for the first ranges algorithm used on a c10::List from
inside that closure. Moving the specializations to right after
ListIterator's own definition in List.h closes the window for good.

Also strengthens List_test.cpp's regression coverage: the existing
random_access_iterator static_assert passes on libstdc++/libc++ via
the stdlib's own fallback whether or not the specialization exists, so
it can't actually catch a regression on the CI this repo runs day to
day -- only the 4-hourly windows-arm64-build-test job could, hours
after the fact. Instantiating basic_common_reference<...>::type
directly does distinguish the two: the primary template has no ::type
member, so it's a hard error if the specialization is ever removed,
independent of any stdlib fallback.
